Ignore common temporary files used during editing and translating
vim save a backup copy of the file being edited adding a tilde (~)to the
of the filename (eg vim copies readme.txt to readme.txt~ and then open the original for editing).

When saving a .po file in poedit and possibly other «.po» file editors, they compile it as «.mo»,
so we tell git to ignore them.