Fix Prev / Next Location links in Meteogram
Fixes for the Previous / Next Location links in the Meteogram.

Update PlacesCount in Main when the user clicks Apply/OK in the Config -
so that the Metwogram knows when the config has changed.

Also change the cursor shape of the Mouseover for the Previous / Next
Location links, depending on whether the links are visible or not.